+++ src/app/src/main/kotlin/org/imperial/mrc/hint/Extensions.kt
@@ -18,10 +18,10 @@ import org.springframework.http.ResponseEntity
 import org.springframework.util.LinkedMultiValueMap
 import org.springframework.util.MultiValueMap
 import org.springframework.web.servlet.mvc.method.annotation.StreamingResponseBody
-import java.io.ByteArrayInputStream
-import java.io.IOException
-import java.io.InputStream
-import java.io.OutputStream
+import java.io.*
+import java.security.DigestInputStream
+import java.security.MessageDigest
+import javax.xml.bind.DatatypeConverter
 
 fun httpStatusFromCode(code: Int): HttpStatus
 {
@@ -53,7 +53,7 @@ fun Response.asResponseEntity(): ResponseEntity<String>
     if (this.statusCode == -1)
     {
         return ErrorDetail(httpStatus, "No response returned. The request may have timed out.")
-                .toResponseEntity() as ResponseEntity<String>
+                .toResponseEntity<String>()
     }
 
     return try
@@ -83,7 +83,7 @@ fun Response.asResponseEntity(): ResponseEntity<String>
     {
         logger.error(e.message)
         ErrorDetail(HttpStatus.INTERNAL_SERVER_ERROR, "Could not parse response.")
-                .toResponseEntity() as ResponseEntity<String>
+                .toResponseEntity<String>()
     }
 }
 
@@ -105,7 +105,7 @@ fun formatADRResponse(json: JsonNode): ResponseEntity<String>
         ErrorDetail(HttpStatus.INTERNAL_SERVER_ERROR,
                 json["error"]["message"].asText(),
                 "ADR_ERROR")
-                .toResponseEntity() as ResponseEntity<String>
+                .toResponseEntity<String>()
     }
 
 }
@@ -131,3 +131,13 @@ fun Request.getStreamingResponseEntity(headRequest: (url: String, parameters: Pa
     return ResponseEntity(responseBody, headers, httpStatus)
 }
 
+fun File.md5sum(): String
+{
+    val md = MessageDigest.getInstance("MD5")
+    FileInputStream(this).use { fis ->
+        DigestInputStream(fis, md).use { dis ->
+            dis.readBytes()
+        }
+    }
+    return DatatypeConverter.printHexBinary(md.digest())
+}