+++ src/app/static/src/app/components/adr/SelectDataset.vue
@@ -77,7 +77,7 @@
     import {Language} from "../../store/translations/locales";
     import Vue from "vue";
     import TreeSelect from "@riophae/vue-treeselect";
-    import {mapActionByName, mapMutationByName, mapStateProp} from "../../utils";
+    import {findResource, mapActionByName, mapMutationByName, mapStateProp} from "../../utils";
     import {RootState} from "../../root";
     import Modal from "../Modal.vue";
     import {BaselineMutation} from "../../store/baseline/mutations";
@@ -91,6 +91,7 @@
     } from "../../types";
     import {InfoIcon} from "vue-feather-icons";
     import {VTooltip} from "v-tooltip";
+    import {ADRState} from "../../store/adr/adr";
 
     interface Methods {
         setDataset: (dataset: Dataset) => void;
@@ -102,10 +103,6 @@
         importSurvey: (url: string) => Promise<void>;
         importProgram: (url: string) => Promise<void>;
         importANC: (url: string) => Promise<void>;
-        findResource: (
-            datasetWithResources: any,
-            resourceType: string
-        ) => DatasetResource | null;
         refresh: () => void;
         refreshDatasetMetadata: () => void;
         markResourcesUpdated: () => void;
@@ -144,6 +141,8 @@
         anc: "ANC",
     };
 
+    const namespace = "adr";
+
     export default Vue.extend<Data, Methods, Computed, unknown>({
         data() {
             return {
@@ -160,21 +159,21 @@
                 "baseline",
                 (state: BaselineState) => !!state.shape
             ),
-            schemas: mapStateProp<RootState, ADRSchemas>(
-                null,
-                (state: RootState) => state.adrSchemas!
+            schemas: mapStateProp<ADRState, ADRSchemas>(
+                namespace,
+                (state: ADRState) => state.schemas!
             ),
             selectedDataset: mapStateProp<BaselineState, Dataset | null>(
                 "baseline",
                 (state: BaselineState) => state.selectedDataset
             ),
-            datasets: mapStateProp<RootState, any[]>(
-                null,
-                (state: RootState) => state.adrDatasets
+            datasets: mapStateProp<ADRState, any[]>(
+                namespace,
+                (state: ADRState) => state.datasets
             ),
-            fetchingDatasets: mapStateProp<RootState, boolean>(
-                null,
-                (state: RootState) => state.adrFetchingDatasets
+            fetchingDatasets: mapStateProp<ADRState, boolean>(
+                namespace,
+                (state: ADRState) => state.fetchingDatasets
             ),
             datasetOptions() {
                 return this.datasets.map((d) => ({
@@ -194,12 +193,12 @@
                     title: fullMetaData.title,
                     url: `${this.schemas.baseUrl}${fullMetaData.type}/${fullMetaData.name}`,
                     resources: {
-                        pjnz: this.findResource(fullMetaData, this.schemas.pjnz),
-                        shape: this.findResource(fullMetaData, this.schemas.shape),
-                        pop: this.findResource(fullMetaData, this.schemas.population),
-                        survey: this.findResource(fullMetaData, this.schemas.survey),
-                        program: this.findResource(fullMetaData, this.schemas.programme),
-                        anc: this.findResource(fullMetaData, this.schemas.anc)
+                        pjnz: findResource(fullMetaData, this.schemas.pjnz),
+                        shape: findResource(fullMetaData, this.schemas.shape),
+                        pop: findResource(fullMetaData, this.schemas.population),
+                        survey: findResource(fullMetaData, this.schemas.survey),
+                        program: findResource(fullMetaData, this.schemas.programme),
+                        anc: findResource(fullMetaData, this.schemas.anc)
                     }
                 }
             },
@@ -248,15 +247,6 @@
             importSurvey: mapActionByName("surveyAndProgram", "importSurvey"),
             importProgram: mapActionByName("surveyAndProgram", "importProgram"),
             importANC: mapActionByName("surveyAndProgram", "importANC"),
-            findResource(datasetWithResources: any, resourceType: string) {
-                const metadata = datasetWithResources.resources.find((r: any) => r.resource_type == resourceType);
-                return metadata ? {
-                    url: metadata.url,
-                    lastModified: metadata.last_modified,
-                    metadataModified: metadata.metadata_modified,
-                    outOfDate: false
-                } : null
-            },
             async importDataset() {
                 if (this.newDatasetId) {
                     this.loading = true;