+++ src/app/static/src/app/hintVersion.ts
@@ -1,3 +1,3 @@
 
-export const currentHintVersion = "1.16.1";
 
+export const currentHintVersion = "1.20.0";