+++ src/app/static/src/tests/components/adr/selectDataset.test.ts
@@ -11,11 +11,11 @@ import {BaselineMutation} from "../../../app/store/baseline/mutations";
 import {BaselineActions} from "../../../app/store/baseline/actions";
 import {SurveyAndProgramActions} from "../../../app/store/surveyAndProgram/actions";
 import {ADRSchemas} from "../../../app/types";
-import {RootState} from "../../../app/root";
 import {InfoIcon} from "vue-feather-icons";
 import Mock = jest.Mock;
 import registerTranslations from "../../../app/store/translations/registerTranslations";
 import {expectTranslated} from "../../testHelpers";
+import {ADRState} from "../../../app/store/adr/adr";
 
 describe("select dataset", () => {
 
@@ -26,40 +26,48 @@ describe("select dataset", () => {
         pjnz: "pjnz",
         population: "pop",
         shape: "shape",
-        survey: "survey"
+        survey: "survey",
+        outputZip: "zip",
+        outputSummary: "summary"
     }
 
     const pjnz = {
+        id: "1",
         resource_type: schemas.pjnz,
         url: "pjnz.pjnz",
         last_modified: "2020-11-01",
         metadata_modified: "2020-11-02"
     }
     const shape = {
+        id: "2",
         resource_type: schemas.shape,
         url: "shape.geojson",
         last_modified: "2020-11-03",
         metadata_modified: "2020-11-04"
     }
     const pop = {
+        id: "3",
         resource_type: schemas.population,
         url: "pop.csv",
         last_modified: "2020-11-05",
         metadata_modified: "2020-11-06"
     }
     const survey = {
+        id: "4",
         resource_type: schemas.survey,
         url: "survey.csv",
         last_modified: "2020-11-07",
         metadata_modified: "2020-11-08"
     }
     const program = {
+        id: "5",
         resource_type: schemas.programme,
         url: "program.csv",
         last_modified: "2020-11-07",
         metadata_modified: "2020-11-08"
     }
     const anc = {
+        id: "6",
         resource_type: schemas.anc,
         url: "anc.csv",
         last_modified: "2020-11-09",
@@ -114,6 +122,7 @@ describe("select dataset", () => {
             pop: null,
             survey: null,
             shape: mockDatasetResource({
+                id: "2",
                 url: "shape.geojson",
                 lastModified: "2020-11-03",
                 metadataModified: "2020-11-04"
@@ -138,14 +147,18 @@ describe("select dataset", () => {
         importANC: jest.fn()
     }
 
-    const getStore = (baselineProps: Partial<BaselineState> = {}, rootProps: Partial<RootState> = {}) => {
+    const getStore = (baselineProps: Partial<BaselineState> = {}, adrProps: Partial<ADRState> = {}) => {
         const store = new Vuex.Store({
-            state: mockRootState({
-                adrSchemas: schemas,
-                adrDatasets: fakeRawDatasets,
-                ...rootProps
-            }),
+            state: mockRootState(),
             modules: {
+                adr: {
+                    namespaced: true,
+                    state: {
+                        schemas: schemas,
+                        datasets: fakeRawDatasets,
+                        ...adrProps
+                    }
+                },
                 baseline: {
                     namespaced: true,
                     state: mockBaselineState(baselineProps),
@@ -421,7 +434,7 @@ describe("select dataset", () => {
     });
 
     it("shows fetching dataset controls, and disables TreeSelect, when fetching", () => {
-        const store = getStore({}, {adrFetchingDatasets: true});
+        const store = getStore({}, {fetchingDatasets: true});
         const rendered = shallowMount(SelectDataset, {store});
         expect(rendered.find("#fetching-datasets").classes()).toStrictEqual(["visible"]);
         expect(rendered.find(TreeSelect).attributes("disabled")).toBe("true");
@@ -433,7 +446,7 @@ describe("select dataset", () => {
 
     it("sets current dataset", async () => {
         let store = getStore({},
-            {adrDatasets: [{...fakeRawDatasets[0], ...fakeRawDatasets[1], resources: [shape]}]}
+            {datasets: [{...fakeRawDatasets[0], ...fakeRawDatasets[1], resources: [shape]}]}
         )
         const rendered = mount(SelectDataset, {
             store, stubs: ["tree-select"]
@@ -473,8 +486,8 @@ describe("select dataset", () => {
 
     it("imports baseline files if they exist", async () => {
         const store = getStore({}, {
-            adrDatasets: [{...fakeRawDatasets[0], resources: [pjnz, pop, shape]}]
-        })
+            datasets: [{...fakeRawDatasets[0], resources: [pjnz, pop, shape]}]
+        });
         const rendered = mount(SelectDataset, {store, stubs: ["tree-select"]});
         rendered.find("button").trigger("click");
 
@@ -503,8 +516,8 @@ describe("select dataset", () => {
 
     it("does not import baseline file if it doesn't exist", async () => {
         const store = getStore({}, {
-            adrDatasets: [{...fakeRawDatasets[0], resources: [pjnz]}]
-        })
+            datasets: [{...fakeRawDatasets[0], resources: [pjnz]}]
+        });
         const rendered = mount(SelectDataset, {store, stubs: ["tree-select"]});
         rendered.find("button").trigger("click");
 
@@ -534,8 +547,8 @@ describe("select dataset", () => {
 
     it("imports survey and program files if they exist and shape file exists", async () => {
         const store = getStore({}, {
-            adrDatasets: [{...fakeRawDatasets[0], resources: [shape, survey, program, anc]}]
-        })
+            datasets: [{...fakeRawDatasets[0], resources: [shape, survey, program, anc]}]
+        });
         const rendered = mount(SelectDataset, {store, stubs: ["tree-select"]});
         rendered.find("button").trigger("click");
 
@@ -564,8 +577,8 @@ describe("select dataset", () => {
 
     it("does not import survey and program file if it doesn't exist", async () => {
         const store = getStore({}, {
-            adrDatasets: [{...fakeRawDatasets[0], resources: [shape, survey]}]
-        })
+            datasets: [{...fakeRawDatasets[0], resources: [shape, survey]}]
+        });
         const rendered = mount(SelectDataset, {store, stubs: ["tree-select"]});
         rendered.find("button").trigger("click");
 
@@ -594,8 +607,8 @@ describe("select dataset", () => {
 
     it("does not import any survey and program files if shape file doesn't exist", async () => {
         const store = getStore({}, {
-            adrDatasets: [{...fakeRawDatasets[0], resources: [survey, program, anc]}]
-        })
+            datasets: [{...fakeRawDatasets[0], resources: [survey, program, anc]}]
+        });
         const rendered = mount(SelectDataset, {store, stubs: ["tree-select"]});
         rendered.find("button").trigger("click");
 
@@ -627,7 +640,7 @@ describe("select dataset", () => {
                 shape: mockShapeResponse()
             },
             {
-                adrDatasets: [{...fakeRawDatasets[0], resources: [survey, program, anc]}]
+                datasets: [{...fakeRawDatasets[0], resources: [survey, program, anc]}]
             });
 
         const rendered = mount(SelectDataset, {store, stubs: ["tree-select"]});
@@ -666,8 +679,8 @@ describe("select dataset", () => {
 
     it("renders can not save when button is disabled", async () => {
         const store = getStore({}, {
-            adrDatasets: [{...fakeRawDatasets[0], resources: [shape, survey]}]
-        })
+            datasets: [{...fakeRawDatasets[0], resources: [shape, survey]}]
+        });
 
         const rendered = mount(SelectDataset, {store, stubs: ["tree-select"]});
         rendered.find("button").trigger("click");